requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.7.4
google-genai>=1.0.0
streamlit>=1.40.0
deepeval>=1.0.0
pypdf>=4.0.0
python-docx>=1.0.0
tiktoken>=0.7.0
chromadb>=1.0.0
pandas>=2.0.0
